ai4kg/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.12.1",
    "fastapi==0.104.1",
    "neo4j==5.13.0",
    "networkx>=3.0",
    "pandas>=1.3.0",
    "passlib[bcrypt]==1.7.4",
    "pydantic[email]==2.5.0",
    "pydantic-settings>=2.2.1",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "redis==5.0.1",
    "sqlalchemy==2.0.23",
    "uvicorn==0.24.0",
    "requests>=2.31.0",
    "numpy>=1.20.0",
    "colorama>=0.4.6",
    "httpx==0.24.1",
]
